requirements.txt
ollama
faster-whisper
sounddevice
soundfile
numpy
piper-tts
torch
python-dotenv

